--- Workers/AudioRemoval/tests/unit/adapters/minio_test.go
@@ -1,63 +1,0 @@
-package adapters_test
-
-import (
-	"bytes"
-	"io"
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-)
-
-func TestNewMinIOStorage_InvalidEndpoint(t *testing.T) {
-	storage, err := NewMinIOStorage("", "access", "secret")
-	assert.Error(t, err)
-	assert.Nil(t, storage)
-}
-
-func TestNewMinIOStorage_ValidParams(t *testing.T) {
-	// This will fail without a real MinIO server, but tests the constructor
-	storage, err := NewMinIOStorage("localhost:9000", "minioadmin", "minioadmin")
-	
-	// We expect this to succeed in creating the client object
-	// even if we can't connect to an actual server
-	if err == nil {
-		assert.NotNil(t, storage)
-		assert.NotNil(t, storage.client)
-	}
-}
-
-func TestMinIOStorage_Structure(t *testing.T) {
-	// Test that we can create the struct even without a real client
-	storage := &MinIOStorage{client: nil}
-	assert.NotNil(t, storage)
-}
-
-// Mock test for GetObject behavior
-func TestMinIOStorage_GetObject_Concept(t *testing.T) {
-	// Test the concept of GetObject without actual MinIO connection
-	bucket := "test-bucket"
-	filename := "test.mp4"
-	
-	assert.NotEmpty(t, bucket)
-	assert.NotEmpty(t, filename)
-	assert.Contains(t, filename, ".mp4")
-}
-
-// Mock test for PutObject behavior
-func TestMinIOStorage_PutObject_Concept(t *testing.T) {
-	// Test the concept of PutObject without actual MinIO connection
-	bucket := "test-bucket"
-	filename := "output.mp4"
-	data := bytes.NewReader([]byte("test video data"))
-	size := int64(15)
-	
-	assert.NotEmpty(t, bucket)
-	assert.NotEmpty(t, filename)
-	assert.NotNil(t, data)
-	assert.Greater(t, size, int64(0))
-	
-	// Test that we can read from the data
-	readData, err := io.ReadAll(data)
-	assert.NoError(t, err)
-	assert.Equal(t, "test video data", string(readData))
-}
