--- Workers/AudioRemoval/tests/unit/adapters/video_repository_test.go
@@ -1,65 +1,0 @@
-package adapters
-
-import (
-	"audioremoval/internal/domain"
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-)
-
-func TestNewVideoRepository(t *testing.T) {
-	repo := NewVideoRepository()
-	assert.NotNil(t, repo)
-}
-
-func TestVideoRepository_FindByFilename(t *testing.T) {
-	repo := NewVideoRepository()
-	filename := "test-video.mp4"
-	
-	video, err := repo.FindByFilename(filename)
-	
-	assert.NoError(t, err)
-	assert.NotNil(t, video)
-	assert.Equal(t, filename, video.ID)
-	assert.Equal(t, filename, video.Filename)
-	assert.Equal(t, domain.StatusPending, video.Status)
-	assert.False(t, video.CreatedAt.IsZero())
-}
-
-func TestVideoRepository_UpdateStatus(t *testing.T) {
-	repo := NewVideoRepository()
-	
-	tests := []struct {
-		name   string
-		id     string
-		status domain.ProcessingStatus
-	}{
-		{"update to processing", "video-1", domain.StatusProcessing},
-		{"update to completed", "video-2", domain.StatusCompleted},
-		{"update to failed", "video-3", domain.StatusFailed},
-	}
-	
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			err := repo.UpdateStatus(tt.id, tt.status)
-			assert.NoError(t, err)
-		})
-	}
-}
-
-func TestVideoRepository_UpdateStatus_AllStatuses(t *testing.T) {
-	repo := NewVideoRepository()
-	videoID := "test-video"
-	
-	statuses := []domain.ProcessingStatus{
-		domain.StatusPending,
-		domain.StatusProcessing,
-		domain.StatusCompleted,
-		domain.StatusFailed,
-	}
-	
-	for _, status := range statuses {
-		err := repo.UpdateStatus(videoID, status)
-		assert.NoError(t, err)
-	}
-}
